Changes to made to Home and DataScreen
The changes that were made to Home Screen was that when the new expense is entered it will automatically update the pie chart, before in order for the pie chart to update you had to click view data now the pie chart is automatically updated. And doing this a new file was created called LogExpense.js which all the functions were called in there.