Database is connected to the user
The format of the files are changed and made it as a CRUD format where all the files are in a specific path and a specific folder. For the profile screen I added the another file which is the EditProfileScreen so that's where the user will be able to make all the changes like changing username and password which is connected to the database and also the logout function works so when the user logs out the data still gets saved. So the profile works and now you can take a photo and have access to the camera but still work in progress image is not uploading to the avatar container.