Updated the files and edited the code
So I deleted the Log Expense file because there was already a file created called UserContext which was created after I created the LogExpense. And when the UserContext file was created Aaron added the add and delete expenses in that file for users and the database was saved in that file so there were two files which had same  functionalities and it was creating errors in the dataScreen so I deleted the LogExpense file and used the database that was stored in the UserContext to update my DataScreen file now the the dataScreen saves the user data and it fetches data when the user adds anything.